Update node discovery for pod hosts
Update the discovery of pods and nodes using the "service" discovery
mode to only consider the hostIPs of pods that are running. Using the
pods hostIP ensures that only nodes that are running the services
targeted pods are registered as consul services.

This fixes an issue where pods would be registered on all cluster nodes
even if they were running on only a subset of the nodes.